created dtos.py, parser and use_case updated
