+++ src/interface_adapters/presenters/presenter.py
@@ -1,15 +1,14 @@
 from flask import render_template
 
 
-def get_chart_template(ra, dec, oid, candid, logo_path, stats, img_str):
+def get_chart_template(info_dict):
     html = render_template(
         "template.html",
-        ra=ra,
-        dec=dec,
-        oid=oid,
-        candid=candid,
-        logo_path=logo_path,
-        stats=stats,
-        panstarrs_image=img_str,
+        ra=info_dict["ra"],
+        dec=info_dict["dec"],
+        oid=info_dict["oid"],
+        candid=info_dict["candid"],
+        stats=info_dict["stats"],
+        panstarrs_image=info_dict["img_str"],
     )
     return html, 200
